apps/backend/pyproject.toml
[project]
name = "agentops-demo-backend"
version = "0.2.0"
description = "AgentOps Demo — RAG pipeline with Langfuse v3 + Phoenix + OTel observability"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.12",
    "langchain-ollama>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.0",
    "pypdf>=4.0.0",
    "langfuse>=4.0.0",
    "arize-phoenix>=8.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# =============================================================================
# Ruff — Python Linter & Formatter (replaces flake8, isort, black, pycodestyle)
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", ".", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade (modern Python syntax)
    "B",      # flake8-bugbear (common bugs)
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
    "PIE",    # flake8-pie (misc lints)
    "PT",     # flake8-pytest-style
    "RET",    # flake8-return
    "RUF",    # Ruff-specific rules
    "S",      # flake8-bandit (security)
    "T20",    # flake8-print (no print statements)
    "TCH",    # flake8-type-checking (move type-only imports)
    "TID",    # flake8-tidy-imports
    "LOG",    # flake8-logging
    "ASYNC",  # flake8-async
    "DTZ",    # flake8-datetimez (timezone-aware)
    "FLY",    # flynt (f-string conversion)
    "PERF",   # perflint (performance anti-patterns)
]
ignore = [
    "S101",   # assert in code (needed for tests)
    "B008",   # function call in default arg (FastAPI Depends pattern)
    "RET504", # unnecessary assignment before return
]

[tool.ruff.lint.per-file-ignores]
"demo_agent.py" = ["T20"]     # print OK in demo script
"tests/**" = ["S101", "S106"]  # assert + hardcoded passwords OK in tests
"main.py" = ["E402"]           # module-level imports after code (FastAPI pattern)

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# =============================================================================
# mypy — Static Type Checking
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = true
no_implicit_optional = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_ollama.*",
    "langchain_chroma.*",
    "langchain_community.*",
    "langchain_text_splitters.*",
    "chromadb.*",
    "langfuse.*",
    "phoenix.*",
    "arize.*",
    "openinference.*",
]
ignore_missing_imports = true

# =============================================================================
# pytest
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-q",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring running services",
]

# =============================================================================
# coverage
# =============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*", "demo_agent.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = true
fail_under = 50
